Backend/my_project/backend_api_requirements.txt
anyio==4.11.0
arrow==1.4.0
asgiref==3.10.0
attrs==25.4.0
bcrypt==5.0.0
CacheControl==0.14.3
cachetools==6.2.1
certifi==2025.10.5
cffi==2.0.0
channels==4.3.1
channels_redis==4.3.0
charset-normalizer==3.4.4
click==8.3.0
config==0.5.1
cryptography==46.0.3
Django==5.2.7
django-location-field==2.7.3
djangorestframework==3.16.1
fcm-django==2.3.1
firebase_admin==7.1.0
google-api-core==2.27.0
google-auth==2.41.1
google-cloud-core==2.4.3
google-cloud-firestore==2.21.0
google-cloud-storage==3.4.1
google-crc32c==1.7.1
google-resumable-media==2.7.2
googleapis-common-protos==1.71.0
grpcio==1.76.0
grpcio-status==1.76.0
h11==0.16.0
h2==4.3.0
hpack==4.1.0
httpcore==1.0.9
httpx==0.28.1
hyperframe==6.1.0
icalendar==6.3.1
ics==0.7.2
idna==3.11
msgpack==1.1.2
proto-plus==1.26.1
protobuf==6.33.0
psycopg2==2.9.11
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser==2.23
PyJWT==2.10.1
python-dateutil==2.9.0.post0
python-decouple==3.8
pytz==2025.2
recurring-ical-events==3.8.0
redis==7.0.0
requests==2.32.5
rsa==4.9.1
six==1.17.0
sniffio==1.3.1
sqlparse==0.5.3
swapper==1.4.0
TatSu==5.13.2
typing_extensions==4.15.0
tzdata==2025.2
urllib3==2.5.0
x-wr-timezone==2.0.1
